d2i_ECPrivateKey: split public key setting into a helper
If the public key is not part of the ECPrivateKey, it needs to be
computed. Rather than doing this ad hoc inline, use the function
from the ameth that already does this.

If it is present, decode it after checking that its unused bits octet
is zero. Again use the dedicated setter API to honor an eventual
EC_KEY_METHOD.

There remains a gross bit reading the point point conversion form out of
the first octet of the bit string. This will go away in a later commit.

ok jsing