Fix up authority and subject key identifiers in force pubkey mode
Upstream decided that this nonsense was worth an ABI break and added stuff
to the X509_CTX so they could hang the issuer's public key off it so that
they could adjust the key identifiers as needed. Let's avoid that and do
it the slightly less nasty way by updating the AKI and SKI as needed.

We only do this when force pubkey is in place so we don't change the
semantics of the batshit crazy config language that nobody understands.

ok job