Merge testdsa.h and tesrsa.h into speed.c
Having constant arrays in a header is just stupid (whether the constants
are static or not), and most of the contents of these two headers clearly
belongs into a C file. Since the garbage pile that is speed.c was not ugly
enough, merge all of it there, since it is the only consumer.

discussed with jsing

PS: still waiting for that elusive volunteer who reworks libdes's speed.c
into something resembling C code.