Revert NULL,0 -> OPENSSL_FILE,OPENSSL_LINE from r1.78
This wasn't part of the initial proposal and causes issues in curl downstream.
We could pile more hacks on top of this, but at some point this is getting too
silly.

Relatedly, most of the FOOerr() could be removed, although PEMerr(), RSAerr()
and SSLerr() are used by some downstreams and probably not worth patching out.

Discussed with @vszakats in https://github.com/libressl/portable/issues/1154